Allow optional requirements_pre.txt, also requirements.txt is optional.

+++ mordor/bin/install_packages.sh
@@ -6,5 +6,12 @@
 
 source $1/venvs/$2-$3/bin/activate
 pip install --upgrade pip
-pip install -r $1/apps/$2/$3/requirements.txt
+if [ -e $1/apps/$2/$3/requirements_pre.txt ]
+then
+    pip install -r $1/apps/$2/$3/requirements_pre.txt
+fi
+if [ -e $1/apps/$2/$3/requirements.txt ]
+then
+    pip install -r $1/apps/$2/$3/requirements.txt
+fi
 deactivate
